Fix code after code review
