+++ tests/commands/dev/conftest.py
@@ -2,43 +2,7 @@ import pytest
 from unittest import mock
 
 from briefcase.commands import DevCommand
-from briefcase.commands.base import full_kwargs
 from briefcase.config import AppConfig
-
-
-class DummyDevCommand(DevCommand):
-    """
-    A dummy Dev command that doesn't actually do anything.
-    It only serves to track which actions would be performend.
-    """
-    platform = 'tester'
-    output_format = 'dummy'
-    description = 'Dummy dev command'
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, apps=[], **kwargs)
-
-        self.actions = []
-        self.env = dict(a=1, b=2, c=3)
-
-    def install_dev_dependencies(self, app, **kwargs):
-        self.actions.append(('dev_dependencies', app.app_name, kwargs))
-
-    def get_environment(self, app):
-        return self.env
-
-    def run_dev_app(self, app, env, **kwargs):
-        self.actions.append(('run_dev', app.app_name, kwargs, env))
-        self.env.update(env)
-        return full_kwargs({
-            'run_dev_state': app.app_name,
-            'env': env
-        }, kwargs)
-
-
-@pytest.fixture
-def dummy_dev_command(tmp_path):
-    return DummyDevCommand(base_path=tmp_path)
 
 
 @pytest.fixture
